Modernize wrappers for advanced table helpers.
Covers array keys, aggregate keys, and SlotDefinition.
