+++ python/lsst/afw/table/SConscript
@@ -2,15 +2,11 @@
 from lsst.sconsUtils import scripts
 scripts.BasicSConscript.pybind11(
     ['_table',
-     'arrays/arrays',
-     'aggregates/aggregates',
-     'slots',
      'simple/simple',
      'source/source',
      'match/match',
      'ampInfo/ampInfo',
      'exposure/exposure',
-     'idFactory',
      'wcsUtils',
      ],
     extraSrc={
@@ -20,16 +16,16 @@ scripts.BasicSConscript.pybind11(
                    '_schemaMapper.cc',
                    '_baseColumnView.cc',
                    '_base.cc',
+                   '_idFactory.cc',
+                   '_arrays.cc',
+                   '_aggregates.cc',
+                   '_slots.cc',
                    ],
-        'arrays/arrays': [],
-        'aggregates/aggregates': [],
-        'slots': [],
         'simple/simple': [],
         'source/source': [],
         'match/match': [],
         'ampInfo/ampInfo': [],
         'exposure/exposure': [],
-        'idFactory': [],
         'wcsUtils': [],
     },
     addUnderscore=False,
