--- python/lsst/afw/table/arrays/arraysContinued.py
@@ -1,35 +1,0 @@
-# This file is part of afw.
-#
-# Developed for the LSST Data Management System.
-# This product includes software developed by the LSST Project
-# (https://www.lsst.org).
-# See the COPYRIGHT file at the top-level directory of this distribution
-# for details of code ownership.
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
-__all__ = ["ArrayKey"]
-
-import numpy as np
-
-from lsst.utils import TemplateMeta
-from . import arrays
-
-
-class ArrayKey(metaclass=TemplateMeta):
-    pass
-
-
-ArrayKey.register(np.float32, arrays.ArrayFKey)
-ArrayKey.register(np.float64, arrays.ArrayDKey)
