+++ python/lsst/afw/table/wcsUtils.cc
@@ -50,7 +50,6 @@ void declareUpdateSourceCoords(py::module &mod) {
 }
 
 PYBIND11_MODULE(wcsUtils, mod) {
-    py::module::import("lsst.afw.table.aggregates");
     py::module::import("lsst.afw.table.simple");
     py::module::import("lsst.afw.table.source");
